+++ Assets/Scripts/World/Chunk.cs
@@ -7,18 +7,18 @@ namespace Starfire.Generation
 {
   public interface IChunk
   {
-    int ChunkIndex { get; }
+    long ChunkIndex { get; }
     string ChunkGroupIndex { get; }
     Vector2Int ChunkKey { get; }
   }
 
   public class Chunk : IChunk
   {
-    public int ChunkIndex { get; private set; }
+    public long ChunkIndex { get; private set; }
     public string ChunkGroupIndex { get; private set; }
     public Vector2Int ChunkKey { get; private set; }
 
-    public Chunk(int chunkIndex, Vector2Int key)
+    public Chunk(long chunkIndex, Vector2Int key)
     {
       ChunkIndex = chunkIndex;
       ChunkGroupIndex = ChunkUtils.GetChunkGroupIndex(key);
